refactor: move widgets to general
