+++ lib/routes/setup/setup_accounts_page.dart
@@ -4,8 +4,8 @@ import 'package:flow/l10n/extensions.dart';
 import 'package:flow/objectbox.dart';
 import 'package:flow/objectbox/objectbox.g.dart';
 import 'package:flow/prefs.dart';
-import 'package:flow/widgets/button.dart';
-import 'package:flow/widgets/info_text.dart';
+import 'package:flow/widgets/general/button.dart';
+import 'package:flow/widgets/general/info_text.dart';
 import 'package:flow/widgets/setup/accounts/account_preset_card.dart';
 import 'package:flow/widgets/setup/accounts/add_account_card.dart';
 import 'package:flow/widgets/setup/setup_header.dart';
