+++ lib/routes/setup/setup_categories_page.dart
@@ -6,9 +6,9 @@ import 'package:flow/objectbox/objectbox.g.dart';
 import 'package:flow/utils/utils.dart';
 import 'package:flow/utils/value_or.dart';
 import 'package:flow/widgets/add_category_card.dart';
-import 'package:flow/widgets/button.dart';
+import 'package:flow/widgets/general/button.dart';
 import 'package:flow/widgets/category_card.dart';
-import 'package:flow/widgets/info_text.dart';
+import 'package:flow/widgets/general/info_text.dart';
 
 import 'package:flow/widgets/setup/categories/category_preset_card.dart';
 import 'package:flow/widgets/setup/setup_header.dart';
