+++ lib/routes/setup/setup_profile_picture_page.dart
@@ -5,8 +5,8 @@ import 'dart:ui' as ui;
 import 'package:flow/l10n/extensions.dart';
 import 'package:flow/objectbox.dart';
 import 'package:flow/utils/utils.dart';
-import 'package:flow/widgets/button.dart';
-import 'package:flow/widgets/profile_picture.dart';
+import 'package:flow/widgets/general/button.dart';
+import 'package:flow/widgets/general/profile_picture.dart';
 import 'package:flow/widgets/setup/setup_header.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
