+++ lib/routes/setup_page.dart
@@ -1,6 +1,6 @@
 import 'package:flow/l10n/extensions.dart';
 import 'package:flow/theme/theme.dart';
-import 'package:flow/widgets/button.dart';
+import 'package:flow/widgets/general/button.dart';
 import 'package:flow/widgets/setup/foss_slide.dart';
 import 'package:flow/widgets/setup/offline_slide.dart';
 import 'package:flow/widgets/setup/welcome_slide.dart';
