+++ lib/widgets/export/export_success.dart
@@ -5,9 +5,9 @@ import 'package:flow/sync/export/mode.dart';
 import 'package:flow/theme/theme.dart';
 import 'package:flow/utils/toast.dart';
 import 'package:flow/utils/utils.dart';
-import 'package:flow/widgets/button.dart';
+import 'package:flow/widgets/general/button.dart';
 import 'package:flow/widgets/general/flow_icon.dart';
-import 'package:flow/widgets/info_text.dart';
+import 'package:flow/widgets/general/info_text.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
