+++ lib/widgets/home/greetings_bar.dart
@@ -2,7 +2,7 @@ import 'package:flow/entity/profile.dart';
 import 'package:flow/l10n/extensions.dart';
 import 'package:flow/objectbox.dart';
 import 'package:flow/objectbox/objectbox.g.dart';
-import 'package:flow/widgets/profile_picture.dart';
+import 'package:flow/widgets/general/profile_picture.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
