+++ lib/widgets/home/home/account/no_accounts.dart
@@ -1,7 +1,7 @@
 import 'package:flow/data/flow_icon.dart';
 import 'package:flow/l10n/extensions.dart';
 import 'package:flow/theme/theme.dart';
-import 'package:flow/widgets/button.dart';
+import 'package:flow/widgets/general/button.dart';
 import 'package:flow/widgets/general/flow_icon.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
